+++ cpp/arduino/Godmode.h
@@ -34,6 +34,17 @@ unsigned long micros();
   #endif
 #endif
 
+// different EEPROM implementations have different macros that leak out
+#if !defined(EEPROM_SIZE) && defined(E2END) && (E2END)
+  // public value indicates that feature is available
+  #define EEPROM_SIZE (E2END + 1)
+  // local array size
+  #define _EEPROM_SIZE EEPROM_SIZE
+#else
+  // feature is not available but we want to have the array so other code compiles
+  #define _EEPROM_SIZE (0)
+#endif
+
 class GodmodeState {
   private:
     struct PortDef {
@@ -60,6 +71,7 @@ class GodmodeState {
     struct PortDef serialPort[NUM_SERIAL_PORTS];
     struct InterruptDef interrupt[MOCK_PINS_COUNT]; // not sure how to get actual number
     struct PortDef spi;
+    uint8_t eeprom[_EEPROM_SIZE];
 
     void resetPins() {
       for (int i = 0; i < MOCK_PINS_COUNT; ++i) {
@@ -99,6 +111,12 @@ class GodmodeState {
       }
     }
 
+    void resetEEPROM() {
+      for(int i = 0; i < EEPROM_SIZE; ++i) {
+        eeprom[i] = 255;
+      }
+    }
+
     void reset() {
       resetClock();
       resetPins();
@@ -106,6 +124,7 @@ class GodmodeState {
       resetPorts();
       resetSPI();
       resetMmapPorts();
+      resetEEPROM();
       seed = 1;
     }